gen-ai/orchestrator-server/src/main/python/server/pyproject.toml
[tool.poetry]
name = "gen_ai_orchestrator"
version = "0.1.0"
description = ""
authors = ["Mohamed ASSOUKTI <mohamed.assoukti@partnre.com>"]
readme = "README.md"
packages = [{include = "gen_ai_orchestrator", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.25.0"
pydantic-settings="^2.1.0"
fastapi = "^0.109.2"
langchain = "~0.2.0"
langchain-openai = "^0.1"
tiktoken = "^0"
opensearch-py = "^2.4.2"
path = "^16.9.0"
colorlog = "^6.8.0"
boto3 = "^1.34.25"
urllib3 = "^1.26"
jinja2 = "^3.1.3"
langfuse = "^2.33.0"
httpx-auth-awssigv4 = "^0.1.4"
python-multipart = "^0.0.9"
pillow = "^10.4.0"
langchain-google-vertexai = ">=1.0.10"
pdf2image = "^1.17.0"
certifi = "2024.7.4"
langchain-community = "^0.2.16"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
pip-audit = "^2.7.0"
bandit = {version = "^1.7.7", extras = ["json"]}
cyclonedx-bom = "^4.1.4"

[tool.poetry.group.test.dependencies]
tox = "^4.11.4"
coverage = "^7.4.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.6"

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"